pyproject.toml
[build-system]
requires = ['setuptools >= 61.0.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'hipsta'
description = 'A python package for hippocampal shape and thickness analysis'
license = {file = 'LICENSE'}
requires-python = '>=3.9'
authors = [
    {name = 'Kersten Diers', email = 'kersten.diers@dzne.de'},
    {name = 'Martin Reuter', email = 'martin.reuter@dzne.de'}
]
maintainers = [
    {name = 'Kersten Diers', email = 'kersten.diers@dzne.de'},
    {name = 'Martin Reuter', email = 'martin.reuter@dzne.de'}
]
keywords = [
    'hippocampus',
    'shape',
    'thickness',
    'geometry',
]
classifiers = [
    'Operating System :: Unix',
    'Operating System :: MacOS',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Natural Language :: English',
    'License :: OSI Approved :: MIT License',
    'Intended Audience :: Science/Research',
]
dynamic = ["version", "readme", "dependencies"]

[project.optional-dependencies]
build = [
    'build',
    'twine',
]
doc = [
    'furo!=2023.8.17',
    'matplotlib',
    'memory-profiler',
    'numpydoc',
    'sphinx!=7.2.*',
    'sphinxcontrib-bibtex',
    'sphinx-copybutton',
    'sphinx-design',
    'sphinx-gallery',
    'sphinx-issues',
    'myst-parser',
    'pypandoc',
    'nbsphinx',
    'IPython', # For syntax highlighting in notebooks
    'ipykernel',
]
style = [
    'bibclean',
    'codespell',
    'pydocstyle[toml]',
    'ruff',
]
test = [
    'pytest',
    'pytest-cov',
    'pytest-timeout',
]
all = [
    'hipsta[doc]',
    'hipsta[build]',
    'hipsta[style]',
]
full = [
    'hipsta[all]',
]

[project.urls]
homepage = 'https://github.com/Deep-MI/Hipsta'
documentation = 'https://github.com/Deep-MI/Hipsta'
source = 'https://github.com/Deep-MI/Hipsta'
tracker = 'https://github.com/Deep-MI/Hipsta/issues'

[project.scripts]
run_hipsta = 'hipsta.cli:main'

[tool.setuptools.dynamic]
version = {file = 'VERSION'}
readme = {file = 'README.md', content-type = "text/markdown"}
dependencies = {file = 'requirements.txt'}

[tool.setuptools.packages.find]
include = ['hipsta*']
exclude = ['docker', 'singularity']

[tool.setuptools.package-data]
"hipsta.doc" = ["*.md"]

[tool.pydocstyle]
convention = 'numpy'
ignore-decorators = '(copy_doc|property|.*setter|.*getter|pyqtSlot|Slot)'
match = '^(?!setup|__init__|test_).*\.py'
match-dir = '^fsqc.*'
add_ignore = 'D100,D104,D107'

[tool.ruff]
line-length = 120
extend-exclude = [
    ".github",
    "doc",
    "docker",
    "images",
    "singularity",
]

[tool.ruff.lint]
# https://docs.astral.sh/ruff/linter/#rule-selection
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    # "SIM",
    # isort
    "I",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"hipsta/computeThickness.py" = ["E501"]   # long lines for easier readability 

[tool.pytest.ini_options]
minversion = '6.0'
filterwarnings = []
addopts = [
    "--import-mode=importlib",
#    "--junit-xml=junit-results.xml",
#    "--durations=20",
#    "--verbose",
]
